requirements.txt
python-telegram-bot==20.3
google-api-python-client==2.95.0
google-auth==2.23.4
nest-asyncio
